Switch to `asdf(1)` part 2
I'm toying more and more with the idea of using `gvim(1)` as my main
editor, performance of a GUI editor is much smoother than terminal
`vim(1)`. With dispatch.vim getting `vim(1) 8's job support, and using
the builtin `:terminal` for `:Start` and `:Spawn` tasks, I'll be able to
develop as I'm used to.

A downside of `chruby(1)` is that it modifies your `$PATH`, meaning you
have to first `cd` into the directory in your terminal, then open
`gvim(1)` to get the `$PATH` set correctly. While I love `chruby(1)` for
its simplicity, the overhead of using a terminal to open a vim instance
just doesn't seem worth it.

Switching to `asdf(1)` is a set towards a full `gvim(1)` future.