Add theming support
This commit adds theming support with the `theme` command, it can toggle
between light and dark, or reload the configuration. Once a theme
changes, we reload the configuration for Kitty and `tmux(1)`.

To be able to control Kitty from inside of `tmux(1)`, we have to create
a UNIX socket that Kitty communicates via. To be able to set the
`--listen-on` configuration option in Kitty when opening the GUI
application, we configure the command-line options in
`macos-launch-services-cmdline` as per
https://sw.kovidgoyal.net/kitty/faq.html#id7

To ease switching between the themes, we extract each of the themes into
their own configuration files for Kitty and `tmux(1)`.

To get the `tmux(1)` colorscheme to be reloadable - we force `tmux(1)`
to track `KITTY_THEME` upon calling `theme NAME_OF_THEME`.

References #28