Fix vim-plugins script
The `git stash` commands failed to push the dirty tree onto the stash,
so the command silently failed on the first vim plugin and halted the
rest.

Instead, we are only pushing the dirty tree onto the stash when the
tree is dirty and only popping said stash if it was pushed in the first
place.